# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: FinanceOS Enterprise Finance AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /variance:
    get:
      operationId: listVarianceDrivers
      tags: [variance]
      summary: List variance drivers
      parameters:
        - name: metricId
          in: query
          schema:
            type: string
            format: uuid
        - name: period
          in: query
          schema:
            $ref: "#/components/schemas/Period"
        - name: fiscalYear
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of variance drivers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VarianceDriverListResponse"
  /variance/forecast:
    get:
      operationId: listForecasts
      tags: [variance]
      summary: List forecast vs actuals
      parameters:
        - name: period
          in: query
          schema:
            $ref: "#/components/schemas/Period"
        - name: fiscalYear
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Forecast records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastListResponse"

  # ── Close ─────────────────────────────────────────────────────
...
  /governance/approvals:
    get:
      operationId: listApprovals
      tags: [governance]
      summary: List approval requests
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ApprovalStatus"
        - name: resourceType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Approval requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalListResponse"
  /governance/approvals/{id}/decide:
    post:
      operationId: decideApproval
      tags: [governance]
      summary: Approve or reject a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
  /governance/audit:
    get:
      operationId: listAuditEvents
      tags: [governance]
      summary: List audit events
      parameters:
        - name: actorId
          in: query
          schema:
            type: string
            format: uuid
        - name: resourceType
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventListResponse"
...
  /evals/runs:
    get:
      operationId: listEvalRuns
      tags: [evals]
      summary: List evaluation runs
      parameters:
        - name: suiteId
          in: query
          schema:
            type: string
            format: uuid
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Eval runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalRunListResponse"
    post:
      operationId: createEvalRun
      tags: [evals]
      summary: Start a new evaluation run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvalRunRequest"
      responses:
        "201":
          description: Created eval run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalRun"

  # ── Agents ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Period:
      type: string
      enum: [Q1, Q2, Q3, Q4, FY, MTD, YTD, LTM]
    Currency:
      type: string
      enum: [USD, EUR, GBP, JPY, CAD]
...
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected, escalated, expired]

    # ── Finance Models ───────────────────────────────────────
...
    VarianceDriver:
      type: object
      properties:
        id:
          type: string
          format: uuid
        metricId:
          type: string
          format: uuid
        driver:
          type: string
        impact:
          type: number
        impactPct:
          type: number
        explanation:
          type: string
        category:
          type: string
          enum: [volume, price, mix, fx, one-time, structural]
        period:
          $ref: "#/components/schemas/Period"
        fiscalYear:
          type: integer
      required:
        - id
        - metricId
        - driver
        - impact
        - impactPct
        - explanation
        - category
        - period
        - fiscalYear
    VarianceDriverListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/VarianceDriver"
        total:
          type: integer
      required:
        - data
        - total
    Forecast:
      type: object
      properties:
        id:
          type: string
          format: uuid
        metricId:
          type: string
          format: uuid
        metricName:
          type: string
        period:
          $ref: "#/components/schemas/Period"
        fiscalYear:
          type: integer
        actual:
          type: number
          nullable: true
        budget:
          type: number
        forecast:
          type: number
        prior:
          type: number
          nullable: true
        currency:
          $ref: "#/components/schemas/Currency"
        confidence:
          type: number
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - metricId
        - metricName
        - period
        - fiscalYear
        - budget
        - forecast
        - currency
        - confidence
        - updatedAt
    ForecastListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Forecast"
        total:
          type: integer
      required:
        - data
        - total
...
    ApprovalRequestApprover:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        decidedAt:
          type: string
          format: date-time
          nullable: true
        comment:
          type: string
      required:
        - userId
        - name
        - role
        - status
    ApprovalRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        resourceType:
          type: string
          enum: [metric, document, workflow, close_task, policy]
        resourceId:
          type: string
          format: uuid
        resourceLabel:
          type: string
        requestedBy:
          type: string
          format: uuid
        requesterName:
          type: string
        approvers:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequestApprover"
        requiredApprovals:
          type: integer
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        priority:
          type: string
          enum: [low, normal, high, urgent]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - resourceType
        - resourceId
        - resourceLabel
        - requestedBy
        - requesterName
        - approvers
        - requiredApprovals
        - status
        - priority
        - createdAt
        - updatedAt
    ApprovalListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequest"
        total:
          type: integer
      required:
        - data
        - total
    ApprovalDecisionRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        comment:
          type: string
      required:
        - decision
    AuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actorId:
          type: string
          format: uuid
        actorName:
          type: string
        actorRole:
          type: string
        action:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
          format: uuid
        resourceLabel:
          type: string
        timestamp:
          type: string
          format: date-time
        outcome:
          type: string
          enum: [success, failure, denied]
        details:
          type: object
          additionalProperties: true
      required:
        - id
        - actorId
        - actorName
        - actorRole
        - action
        - resourceType
        - resourceId
        - resourceLabel
        - timestamp
        - outcome
    AuditEventListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        total:
          type: integer
      required:
        - data
        - total
...
    EvalRunResult:
      type: object
      properties:
        caseId:
          type: string
          format: uuid
        caseName:
          type: string
        output:
          type: string
        scores:
          type: object
          additionalProperties:
            type: number
        latencyMs:
          type: integer
        passed:
          type: boolean
      required:
        - caseId
        - caseName
        - output
        - scores
        - latencyMs
        - passed
    EvalRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        suiteId:
          type: string
          format: uuid
        suiteName:
          type: string
        agentId:
          type: string
          format: uuid
        agentName:
          type: string
        status:
          type: string
          enum: [queued, running, complete, failed]
        results:
          type: array
          items:
            $ref: "#/components/schemas/EvalRunResult"
        aggregateScores:
          type: object
          additionalProperties:
            type: number
        passRate:
          type: number
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - suiteId
        - suiteName
        - agentId
        - agentName
        - status
        - results
        - createdAt
    EvalRunListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EvalRun"
        total:
          type: integer
      required:
        - data
        - total
    CreateEvalRunRequest:
      type: object
      properties:
        suiteId:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
      required:
        - suiteId
        - agentId

    # ── Agents ──────────────────────────────────────────────────
...