# lib/api-zod/src/generated/types/approvalRequestApprover.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceOS Enterprise Finance AI Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { ApprovalStatus } from "./approvalStatus";

export interface ApprovalRequestApprover {
  userId: string;
  name: string;
  role: string;
  status: ApprovalStatus;
  decidedAt?: Date | null;
  comment?: string;
}