# lib/api-zod/src/generated/types/evalRunListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinanceOS Enterprise Finance AI Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { EvalRun } from "./evalRun";

export interface EvalRunListResponse {
  data: EvalRun[];
  total: number;
}